refactor (be): repackage codes by feature
