+++ server/src/main/java/com/foolproof/server/user/UserApiController.java
@@ -1,4 +1,4 @@
-package com.foolproof.server.controller;
+package com.foolproof.server.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,10 +6,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.foolproof.server.domain.User;
-import com.foolproof.server.dto.UserDTO;
-import com.foolproof.server.service.UserService;
 
 import lombok.RequiredArgsConstructor;
