+++ server/src/main/java/com/foolproof/server/user/UserRepository.java
@@ -1,7 +1,6 @@
-package com.foolproof.server.repository;
+package com.foolproof.server.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.foolproof.server.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
