+++ server/src/main/java/com/foolproof/server/user/UserService.java
@@ -1,11 +1,6 @@
-package com.foolproof.server.service;
+package com.foolproof.server.user;
 
 import org.springframework.stereotype.Service;
-
-import com.foolproof.server.domain.User;
-import com.foolproof.server.dto.UserDTO;
-import com.foolproof.server.repository.UserRepository;
-
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
